tests: restructure integration tests
