+++ tests/_data/sbom_demo-results/bare/local-dependencies_from-setup.snap.json
@@ -144,9 +144,9 @@
     },
     {
       "type": "library",
-      "name": "my-local-b",
+      "name": "my-local-b-off",
       "version": "0.0.0",
-      "bom-ref": "my-local-b@0.0.0",
+      "bom-ref": "my-local-b-off@0.0.0",
       "description": "demo: my-local-b-off - a package with a different name than its dir",
       "licenses": [
         {
@@ -156,7 +156,7 @@
           }
         }
       ],
-      "purl": "pkg:npm/my-local-b@0.0.0?vcs_url=git%2Bssh%3A//git%40github.com/CycloneDX/cyclonedx-node-npm.git#demo/local-dependencies/project/packages/my-local-b",
+      "purl": "pkg:npm/my-local-b-off@0.0.0?vcs_url=git%2Bssh%3A//git%40github.com/CycloneDX/cyclonedx-node-npm.git#demo/local-dependencies/project/packages/my-local-b",
       "externalReferences": [
         {
           "url": "https://github.com/CycloneDX/cyclonedx-node-npm/issues",
@@ -240,11 +240,11 @@
     {
       "ref": "my-local-a@0.0.0",
       "dependsOn": [
-        "my-local-b@0.0.0"
+        "my-local-b-off@0.0.0"
       ]
     },
     {
-      "ref": "my-local-b@0.0.0"
+      "ref": "my-local-b-off@0.0.0"
     },
     {
       "ref": "my-noname@0.0.0"
