+++ tests/_data/sbom_demo-results/bare/local-dependencies_from-setup.snap.xml
@@ -105,8 +105,8 @@
         <property name="cdx:npm:package:private">true</property>
       </properties>
     </component>
-    <component type="library" bom-ref="my-local-b@0.0.0">
-      <name>my-local-b</name>
+    <component type="library" bom-ref="my-local-b-off@0.0.0">
+      <name>my-local-b-off</name>
       <version>0.0.0</version>
       <description>demo: my-local-b-off - a package with a different name than its dir</description>
       <licenses>
@@ -114,7 +114,7 @@
           <id>Apache-2.0</id>
         </license>
       </licenses>
-      <purl>pkg:npm/my-local-b@0.0.0?vcs_url=git%2Bssh%3A//git%40github.com/CycloneDX/cyclonedx-node-npm.git#demo/local-dependencies/project/packages/my-local-b</purl>
+      <purl>pkg:npm/my-local-b-off@0.0.0?vcs_url=git%2Bssh%3A//git%40github.com/CycloneDX/cyclonedx-node-npm.git#demo/local-dependencies/project/packages/my-local-b</purl>
       <externalReferences>
         <reference type="issue-tracker">
           <url>https://github.com/CycloneDX/cyclonedx-node-npm/issues</url>
@@ -170,9 +170,9 @@
       <dependency ref="my-noname@0.0.0"/>
     </dependency>
     <dependency ref="my-local-a@0.0.0">
-      <dependency ref="my-local-b@0.0.0"/>
+      <dependency ref="my-local-b-off@0.0.0"/>
     </dependency>
-    <dependency ref="my-local-b@0.0.0"/>
+    <dependency ref="my-local-b-off@0.0.0"/>
     <dependency ref="my-noname@0.0.0"/>
   </dependencies>
 </bom>
